aws/text_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
aws-cdk.asset-awscli-v1==2.2.240
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-codestar-alpha==2.202.0a0
aws-cdk.cloud-assembly-schema==44.8.0
aws-cdk-lib==2.202.0
boto3==1.38.42
boto3-stubs==1.38.42
botocore==1.38.42
cattrs==24.1.3
certifi==2025.6.15
charset-normalizer==3.4.2
constructs==10.4.2
dataclasses-json==0.6.7
faiss-cpu==1.11.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.112.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-aws==0.2.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
marshmallow==3.26.1
multidict==6.5.1
mypy_extensions==1.1.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pip==25.1
propcache==0.3.2
publication==0.0.3
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typeguard==2.13.3
typing_extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.3.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
